Nullable arguments in OAuthException constructor
At lines 23 and 25, we check if these parameters are not null. Checking them doesn't make much sense if type string is required.

This change will also fix an issue: in src/Security/Firewall/ConnectAuthenticationListener.php at line 65, we create an OAuthException with $request->query->get('error_description') as a second argument. But this can be bull in the $request query doesn't carry any error_description parameter.